+++ ElvUI_MerathilisUI/Locales/deDE.lua
@@ -390,6 +390,7 @@ L["Embed Settings"] = "Einbettungseinstellungen"
 L["With this option you can embed your Details into an own Panel."] =
 	"Mit dieser Option kannst Du Dein Details in ein eigenes Panel einbetten."
 L["Number of Windows"] = "Anzahl der Fenster"
+L["Window %d"] = "Fenster %d"
 L["Reset Settings"] = "Zurücksetzen"
 L["Toggle Direction"] = "Richtung umschalten"
 L["TOP"] = "OBEN"