+++ ElvUI_MerathilisUI/Locales/enUS.lua
@@ -332,6 +332,7 @@ L["Action Status"] = true
 L["Embed Settings"] = true
 L["With this option you can embed your Details into an own Panel."] = true
 L["Number of Windows"] = true
+L["Window %d"] = true
 L["Reset Settings"] = true
 L["Toggle Direction"] = true
 L["TOP"] = true